Fix argument type from int to uint for signInt16 function

+++ font/woff2.go
@@ -366,8 +366,8 @@ func parseGlyfTransformed(b []byte) ([]byte, []byte, error) {
 				return nil, nil, ErrInvalidFontData
 			}
 
-			signInt16 := func(flag byte, pos int) int16 {
-				if flag&(1<<uint(pos)) != 0 {
+			signInt16 := func(flag byte, pos uint) int16 {
+				if flag&(1<<pos) != 0 {
 					return 1 // positive if bit on position is set
 				}
 				return -1
